refactor(store): extract _SearchEngineMixin to _search.py (#280)
Fifth step of the VstashStore decomposition (#280): the hybrid-search pipeline.
The 13-method contiguous block (_compute_search_cache_key, _resolve_retrieval_mode,
_resolve_rrf_weights, _build_fts_match_query, _fuse_rrf_scores, _apply_recency_boost,
_build_search_results, search, miss_analysis, _build_miss_suggestions, _mmr_dedup,
_get_filter_conditions, _build_doc_filter) moves into _SearchEngineMixin in
vstash/store/_search.py; VstashStore inherits it. Facade __init__.py drops from
~3780 to ~2040 LOC.

Method bodies are byte-identical except two required edits: the explicit
`VstashStore._get_filter_conditions(...)` static calls inside _build_doc_filter
become `_SearchEngineMixin._get_filter_conditions(...)` — the staticmethod now
lives on the mixin, and referencing VstashStore from _search.py would be a
circular import (verified by diff: only those 2 lines differ).

Cross-mixin helpers still on the facade (_stem_terms, _build_idf_cache,
_compute_adaptive_rrf_params, _bump_cache_epoch) resolve via the MRO; they move
to _search.py in the follow-up PR6.

- pytest tests/ -> 1259 passed (1253 + 6 new test_search_engine.py), 6 deselected
- ruff check/format clean
